feat:1217-wip-validation-and-database-connections
