+++ .env.template
@@ -1,3 +1,9 @@
+XPAND_DATABASE__PASSWORD=
+XPAND_DATABASE__HOST=
+XPAND_DATABASE__USER=
+XPAND_DATABASE__DATABASE=
+XPAND_DATABASE__PORT=
+
 DATABASE__PASSWORD=
 DATABASE__HOST=localhost
 DATABASE__USER=sa
