+++ knexfile.js
@@ -4,16 +4,19 @@ require('dotenv').config()
 /**
  * @type { Object.<string, import("knex").Knex.Config> }
  */
+
+//todo: update knex file to support multiple connections
+//todo: fix docker file, read from 2 different knex files?
 module.exports = {
   client: 'mssql',
   dev: {
     client: 'mssql',
     connection: {
-      host: process.env.DATABASE__HOST,
-      database: process.env.DATABASE__DATABASE,
-      user: process.env.DATABASE__USER,
-      password: process.env.DATABASE__PASSWORD,
-      port: parseInt(process.env.DATABASE__PORT),
+      host: process.env.LEASING_DATABASE__HOST,
+      database: process.env.LEASING_DATABASE__DATABASE,
+      user: process.env.LEASING_DATABASE__USER,
+      password: process.env.LEASING_DATABASE__PASSWORD,
+      port: parseInt(process.env.LEASING_DATABASE__PORT),
     },
     migrations: {
       tableName: 'knex_migrations',
