+++ src/services/lease-service/tests/waitingList.test.ts
@@ -4,11 +4,11 @@ import KoaRouter from '@koa/router'
 import bodyParser from 'koa-bodyparser'
 
 import { routes } from '../index'
-import * as xpandSoapAdapter from '../adapters/xpand-soap-adapter'
+import * as xpandSoapAdapter from '../adapters/xpand/xpand-soap-adapter'
 import {
   addApplicantToToWaitingList,
   getWaitingList,
-} from '../adapters/xpand-soap-adapter'
+} from '../adapters/xpand/xpand-soap-adapter'
 import { WaitingList } from '../../../../../onecore-types'
 import * as http from 'http'
 import { HttpStatusCode } from 'axios'
